Updating UI to use auto layout and adding launch images so we can use the full screen realestate for new large screens.

+++ WaterChat/WaterChat/WATPeerManager.m
@@ -288,7 +288,7 @@
 - (UIImage *)randomImage
 {
     NSUInteger randomIndex = arc4random()%10;
-    NSString *imageString = [NSString stringWithFormat:@"avatar%d",randomIndex];
+    NSString *imageString = [NSString stringWithFormat:@"avatar%lu",(unsigned long)randomIndex];
     UIImage *selectedImage = [UIImage imageNamed:imageString];
     return selectedImage;
 }
